feat: add table and util for table view
