+++ tolife2/utils/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class UtilsConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "tolife2.utils"
